refactor: Sections로 리팩터링
